requirements.txt
requests==2.31.0
python-dotenv==1.0.1
APScheduler==3.10.4
anthropic==0.25.0
